requirements.txt
comtypes==1.4.15
python-nmap==0.7.1
python3-nmap==1.9.1
pywifi==1.1.12
scapy==2.7.0
simplejson==3.20.2
pycryptodome==3.23.0